requirements.txt
albumentations==1.1.0
comet_ml==3.20.0
pretrainedmodels==0.7.4
efficientnet_pytorch==0.7.1
# torch==1.10
omegaconf==2.1.1
pytorch_toolbelt==0.4.4
hydra_core==1.1.1
hydra_colorlog==1.1.0
pytest==6.2.5
numpy==1.21.3
pandas==1.3.4
# torchvision==0.10.1
pytorch_lightning==1.5.0
Pillow==8.4.0
scikit_learn==1.0.1
adabound==0.0.5
adabelief-pytorch==0.2.1
ranger-adabelief==0.1.0
timm==0.4.12
torch-optimizer==0.3.0
wandb==0.12.6
einops==0.3.2
pyarrow==6.0.0
rich==10.12.0

